DEFT: move dedicated tests to tests/test_deft.py (review feedback)
Move TestDeftPaRa and TestDeftMerge out of tests/test_custom_models.py into a
new tests/test_deft.py (with a local MLP fixture). The shared identity-at-init
DEFT cases stay in test_custom_models.py's TEST_CASES.
